reorganised the modules
- split internals angle, latitude and longitude to different modules

- Geo now called LatLon.
